requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
bleach==5.0.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.0
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-bootstrap4==22.1
django-cors-headers==3.11.0
django-material==1.10.0
django-rest-framework==0.1.0
django-summernote==0.8.20.0
django-taggit==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
filelock==3.6.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==8.3.2
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycparser==2.21
pygame==2.1.2
PyJWT==2.3.0
pyparsing==3.0.8
python-dateutil==1.5
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.0
webencodings==0.5.1
whitenoise==6.0.0